Moved button to appropriate folder
